Other_Repos/lama/requirements.txt
pyyaml
tqdm
numpy
easydict==1.9.0
scikit-image
scikit-learn
opencv-python
tensorflow
joblib
matplotlib
pandas
albumentations
hydra-core
pytorch-lightning
tabulate
kornia
webdataset
packaging
scikit-learn
wldhx.yadisk-direct
torchvision